requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==1.0.0
Django==3.2.16
django-crispy-forms==1.14.0
django-email-verification==0.3.1
django-filter==22.1
django-heroku==0.3.1
django-recaptcha==3.0.0
django-templated-mail==1.1.1
django-tinymce==3.5.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
fontawesome-free==5.15.4
fontawesomefree==6.2.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.8.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
